requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
python-dateutil>=2.8.0
python-docx>=1.0.0
